handful of enhancements...
* change visualization hue argument to float
* explicitly cast many implicit casts where intended to prevent warnings
* change size and index types for grids to signed integers
* change ara* search state members to signed types
* factor out clamp function to common algorithm header
* silence some precedence warnings
* lots of type deduction replacements

+++ smpl/include/smpl/debug/marker_utils.h
@@ -11,7 +11,7 @@ auto MakeEmptyMarker() -> Marker;
 auto MakeSphereMarker(
     double x, double y, double z,
     double radius,
-    int hue,
+    float hue,
     const std::string& frame_id,
     const std::string& ns,
     int id = 0) -> Marker;
