+++ smpl/src/debug/marker_utils.cpp
@@ -14,7 +14,7 @@ auto MakeEmptyMarker() -> Marker {
 auto MakeSphereMarker(
     double x, double y, double z,
     double radius,
-    int hue,
+    float hue,
     const std::string& frame_id,
     const std::string& ns,
     int id) -> Marker
